# docs/openapi.yaml
openapi: 3.1.0
info:
  title: REKAM API
  version: 0.1.0
  description: Backend API for REKAM local and on-premise deployment.
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...
paths:
...
  /api/auth/login:
    post:
      security: []
      summary: Login and set HttpOnly refresh cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Access token and user profile.
        "401":
          description: Invalid credentials.
  /api/auth/refresh:
    post:
      security: []
      summary: Refresh access token from HttpOnly cookie
      responses:
        "200":
          description: New access token.
        "401":
          description: Refresh token missing, invalid, expired, or revoked.
...
  /api/export/activity/{id}/zip:
    get:
      summary: Export activity media as ZIP
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: ZIP stream. }
  /api/export/activity/{id}/pdf:
    get:
      summary: Export activity report as PDF
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: PDF stream. }
...